Completely merge the separate JSON presenters into the service
