+++ tests/h/services/annotation_json_presentation/service_test.py
@@ -1,7 +1,8 @@
-from unittest.mock import sentinel
+from datetime import datetime
 
 import pytest
 from h_matchers import Any
+from pyramid.authorization import Everyone
 from sqlalchemy import event
 
 from h.security.permissions import Permission
@@ -10,30 +11,114 @@ from h.traversal import AnnotationContext
 
 
 class TestAnnotationJSONPresentationService:
-    def test_present(self, service, BasicJSONPresenter):
-        result = service.present(annotation=sentinel.annotation)
-
-        BasicJSONPresenter.return_value.present.assert_called_once_with(
-            sentinel.annotation
-        )
-        assert result == BasicJSONPresenter.return_value.present.return_value
-
-    def test_present_for_user(
-        self, service, user, annotation, BasicJSONPresenter, flag_service
+    def test_present(
+        self, service, annotation, links_service, user_service, DocumentJSONPresenter
     ):
-        BasicJSONPresenter.return_value.present.return_value = {"presenter": 1}
+        annotation.created = datetime(2016, 2, 24, 18, 3, 25, 768)
+        annotation.updated = datetime(2016, 2, 29, 10, 24, 5, 564)
+        annotation.references = ["referenced-id-1", "referenced-id-2"]
+        annotation.extra = {"extra-1": "foo", "extra-2": "bar"}
 
+        result = service.present(annotation)
+
+        links_service.get_all.assert_called_once_with(annotation)
+        user_service.fetch.assert_called_once_with(annotation.userid)
+
+        assert result == {
+            "id": annotation.id,
+            "created": "2016-02-24T18:03:25.000768+00:00",
+            "updated": "2016-02-29T10:24:05.000564+00:00",
+            "user": annotation.userid,
+            "uri": annotation.target_uri,
+            "group": annotation.groupid,
+            "text": annotation.text,
+            "tags": annotation.tags,
+            "permissions": {
+                "read": [annotation.userid],
+                "admin": [annotation.userid],
+                "update": [annotation.userid],
+                "delete": [annotation.userid],
+            },
+            "target": annotation.target,
+            "document": DocumentJSONPresenter.return_value.asdict.return_value,
+            "links": links_service.get_all.return_value,
+            "references": annotation.references,
+            "extra-1": "foo",
+            "extra-2": "bar",
+            "user_info": {"display_name": user_service.fetch.return_value.display_name},
+        }
+
+        DocumentJSONPresenter.assert_called_once_with(annotation.document)
+        DocumentJSONPresenter.return_value.asdict.assert_called_once_with()
+
+    def test_present_without_references(self, service, annotation):
+        annotation.references = None
+
+        result = service.present(annotation)
+
+        assert "references" not in result
+
+    def test_present_extra_inherits_correctly(self, service, annotation):
+        annotation.extra = {"id": "DIFFERENT"}
+
+        presented = service.present(annotation)
+
+        # We can't override things (we are applied first)
+        assert presented["id"] == annotation.id
+        # And we aren't mutated
+        assert annotation.extra == {"id": "DIFFERENT"}
+
+    @pytest.mark.parametrize(
+        "shared,readable_by,permission_template",
+        (
+            (False, [], "{annotation.userid}"),
+            (True, [], "group:{annotation.groupid}"),
+            (True, [Everyone], "group:__world__"),
+        ),
+    )
+    def test_present_read_permission(
+        self,
+        service,
+        annotation,
+        identity_permits,
+        shared,
+        readable_by,
+        permission_template,
+    ):
+        annotation.shared = shared
+        identity_permits.return_value = readable_by
+
+        presented = service.present(annotation)
+
+        permission = permission_template.format(annotation=annotation)
+        assert presented["permissions"]["read"] == [permission]
+
+    def test_present_skips_the_read_permission_check_if_group_is_world(
+        self, service, annotation, identity_permits
+    ):
+        annotation.shared = True
+        annotation.groupid = "__world__"
+
+        presented = service.present(annotation)
+
+        identity_permits.assert_not_called()
+        assert presented["permissions"]["read"] == ["group:__world__"]
+
+    def test_present_for_user(self, service, user, annotation, flag_service):
         result = service.present_for_user(annotation, user)
 
-        BasicJSONPresenter.return_value.present.assert_called_once_with(annotation)
         flag_service.flagged.assert_called_once_with(user, annotation)
         flag_service.flag_count.assert_called_once_with(annotation)
-        assert result == {
-            "presenter": 1,
-            "hidden": False,
-            "flagged": flag_service.flagged.return_value,
-            "moderation": {"flagCount": flag_service.flag_count.return_value},
-        }
+        assert result == Any.dict.containing(
+            {
+                # At least one thing from normal serialization
+                "id": Any(),
+                # ... and the things this method adds
+                "hidden": False,
+                "flagged": flag_service.flagged.return_value,
+                "moderation": {"flagCount": flag_service.flag_count.return_value},
+            }
+        )
 
     def test_present_for_user_only_shows_moderation_to_moderators(
         self, service, annotation, user, identity_permits, Identity
@@ -77,13 +162,9 @@ class TestAnnotationJSONPresentationService:
 
     @pytest.mark.usefixtures("with_hidden_annotation")
     def test_present_for_user_hidden_shows_everything_to_moderators(
-        self, service, annotation, user, identity_permits, BasicJSONPresenter
+        self, service, annotation, user, identity_permits
     ):
         identity_permits.return_value = True
-        BasicJSONPresenter.return_value.present.return_value = {
-            "text": sentinel.text,
-            "tags": [sentinel.tags],
-        }
 
         result = service.present_for_user(annotation, user)
 
@@ -92,13 +173,7 @@ class TestAnnotationJSONPresentationService:
         assert result["tags"]
 
     def test_present_all_for_user(
-        self,
-        service,
-        annotation,
-        user,
-        BasicJSONPresenter,
-        flag_service,
-        user_service,
+        self, service, annotation, user, flag_service, user_service
     ):
         annotation_ids = [annotation.id]
 
@@ -107,9 +182,10 @@ class TestAnnotationJSONPresentationService:
         flag_service.all_flagged.assert_called_once_with(user, annotation_ids)
         flag_service.flag_counts.assert_called_once_with(annotation_ids)
         user_service.fetch_all.assert_called_once_with([annotation.userid])
-        BasicJSONPresenter.return_value.present.assert_called_once_with(annotation)
+
         assert result == [
-            BasicJSONPresenter.return_value.present.return_value,
+            # A few indicative fields to show we are serializing
+            Any.dict.containing({"id": Any(), "hidden": False})
         ]
 
     @pytest.mark.parametrize("attribute", ("document", "moderation", "group"))
@@ -152,10 +228,10 @@ class TestAnnotationJSONPresentationService:
         return query_counter
 
     @pytest.fixture
-    def service(self, db_session, flag_service, user_service):
+    def service(self, db_session, links_service, flag_service, user_service):
         return AnnotationJSONPresentationService(
             session=db_session,
-            links_service=sentinel.links_service,
+            links_service=links_service,
             flag_service=flag_service,
             user_service=user_service,
         )
@@ -166,7 +242,7 @@ class TestAnnotationJSONPresentationService:
 
     @pytest.fixture
     def annotation(self, factories):
-        return factories.Annotation(moderation=None)
+        return factories.Annotation(moderation=None, groupid="NOT WORLD")
 
     @pytest.fixture
     def with_hidden_annotation(self, annotation, factories):
@@ -181,9 +257,7 @@ class TestAnnotationJSONPresentationService:
         return patch("h.services.annotation_json_presentation.service.identity_permits")
 
     @pytest.fixture(autouse=True)
-    def BasicJSONPresenter(self, patch):
-        BasicJSONPresenter = patch(
-            "h.services.annotation_json_presentation.service.BasicJSONPresenter"
+    def DocumentJSONPresenter(self, patch):
+        return patch(
+            "h.services.annotation_json_presentation.service.DocumentJSONPresenter"
         )
-        BasicJSONPresenter.return_value.present.return_value = {}
-        return BasicJSONPresenter
